requirements_github.txt
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
python-binance>=1.0.19
PyYAML>=6.0
requests>=2.28.0
python-dateutil>=2.8.2
aiohttp>=3.8.0
websockets>=11.0.0
pycryptodome>=3.15.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0